Add files via upload
I have used for loop to iterate numbers 2 to 100. isPrime boolean method checks if a number is prime by dividing the numbers to square root of the given number using (Math.sqrt(n)) function. If it is divisible it is Not Prime number, If its not divisible it is Printed as Prime Number.