Add files via upload
Imported Scanner function to get input from the user. Boolean function is used to check whether the given number is prime. If loop is used to check if the given number is divisible by 2. If condition is True, It is Prime number. If condition is False, It is not a Prime number.